refactor definitions mess
